+++ packages/multistream-select/test/dialer.spec.ts
@@ -9,7 +9,7 @@ import { pEvent } from 'p-event'
 import pTimeout from 'p-timeout'
 import { Uint8ArrayList } from 'uint8arraylist'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
-import * as mss from '../src/index.js'
+import * as mss from '../src/index.ts'
 
 describe('Dialer', () => {
   describe('dialer.select', () => {