+++ packages/config/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.1.32](https://github.com/libp2p/js-libp2p/compare/config-v1.1.31...config-v1.1.32) (2026-05-30)
+
+
+### Dependencies
+
+* upgrade to aegir v48 ([#3495](https://github.com/libp2p/js-libp2p/issues/3495)) ([29797a5](https://github.com/libp2p/js-libp2p/commit/29797a5bbbfc5b93d91fcdc244d7f65e2ce5e76c))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.18 to ^5.1.19
+    * @libp2p/interface bumped from ^3.2.2 to ^3.2.3
+    * @libp2p/keychain bumped from ^6.1.1 to ^6.1.2
+    * @libp2p/logger bumped from ^6.2.7 to ^6.2.8
+
 ## [1.1.31](https://github.com/libp2p/js-libp2p/compare/config-v1.1.30...config-v1.1.31) (2026-05-09)