+++ packages/connection-encrypter-plaintext/CHANGELOG.md
@@ -72,6 +72,22 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [3.0.21](https://github.com/libp2p/js-libp2p/compare/plaintext-v3.0.20...plaintext-v3.0.21) (2026-05-30)
+
+
+### Dependencies
+
+* upgrade to aegir v48 ([#3495](https://github.com/libp2p/js-libp2p/issues/3495)) ([29797a5](https://github.com/libp2p/js-libp2p/commit/29797a5bbbfc5b93d91fcdc244d7f65e2ce5e76c))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.18 to ^5.1.19
+    * @libp2p/interface bumped from ^3.2.2 to ^3.2.3
+    * @libp2p/peer-id bumped from ^6.0.9 to ^6.0.10
+    * @libp2p/utils bumped from ^7.2.1 to ^7.2.2
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.18 to ^5.1.19
+    * @libp2p/logger bumped from ^6.2.7 to ^6.2.8
+
 ## [3.0.20](https://github.com/libp2p/js-libp2p/compare/plaintext-v3.0.19...plaintext-v3.0.20) (2026-05-16)