+++ packages/gossipsub/CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## [16.0.1](https://github.com/libp2p/js-libp2p/compare/gossipsub-v16.0.0...gossipsub-v16.0.1) (2026-05-30)
+
+
+### Dependencies
+
+* upgrade multiformats to 14 and multiaddr to 13.0.3 ([#3526](https://github.com/libp2p/js-libp2p/issues/3526)) ([5b8813a](https://github.com/libp2p/js-libp2p/commit/5b8813abcbca3fd59e6d137f2a42ff53e115cf99))
+* upgrade to aegir v48 ([#3495](https://github.com/libp2p/js-libp2p/issues/3495)) ([29797a5](https://github.com/libp2p/js-libp2p/commit/29797a5bbbfc5b93d91fcdc244d7f65e2ce5e76c))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.18 to ^5.1.19
+    * @libp2p/interface bumped from ^3.2.2 to ^3.2.3
+    * @libp2p/interface-internal bumped from ^3.1.5 to ^3.1.6
+    * @libp2p/peer-id bumped from ^6.0.9 to ^6.0.10
+    * @libp2p/utils bumped from ^7.2.1 to ^7.2.2
+  * devDependencies
+    * @libp2p/floodsub bumped from ^11.0.22 to ^11.0.23
+    * @libp2p/logger bumped from ^6.2.7 to ^6.2.8
+    * @libp2p/peer-store bumped from ^12.0.20 to ^12.0.21
+
+## [16.0.0](https://github.com/libp2p/js-libp2p/compare/gossipsub-v15.0.23...gossipsub-v16.0.0) (2026-05-27)
+
+
+### ⚠ BREAKING CHANGES
+
+* **gossipsub:** apply decodeRpcLimits to control messages ([#3520](https://github.com/libp2p/js-libp2p/issues/3520))
+
+### Bug Fixes
+
+* **gossipsub:** apply decodeRpcLimits to control messages ([#3520](https://github.com/libp2p/js-libp2p/issues/3520)) ([773dd80](https://github.com/libp2p/js-libp2p/commit/773dd80ded24dbd6b19e675c89fd2f3b45f2d899))
+
+
+### Documentation
+
+* @libp2p/gossipsub ([#3516](https://github.com/libp2p/js-libp2p/issues/3516)) ([c6329a0](https://github.com/libp2p/js-libp2p/commit/c6329a0efdb33f63534468fd90e6bc274ae6fe15))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/floodsub bumped from ^11.0.21 to ^11.0.22
+
 ## [15.0.23](https://github.com/libp2p/js-libp2p/compare/gossipsub-v15.0.22...gossipsub-v15.0.23) (2026-05-16)