+++ packages/interface-compliance-tests/CHANGELOG.md
@@ -75,6 +75,40 @@
   * dependencies
     * @libp2p/utils bumped from ^5.2.4 to ^5.2.5
 
+## [7.0.24](https://github.com/libp2p/js-libp2p/compare/interface-compliance-tests-v7.0.23...interface-compliance-tests-v7.0.24) (2026-05-30)
+
+
+### Bug Fixes
+
+* **utils:** abort message stream when sendData throws during drain flush ([#3527](https://github.com/libp2p/js-libp2p/issues/3527)) ([160a245](https://github.com/libp2p/js-libp2p/commit/160a245858b6ddc052213f10de463e8ddb9a1b76))
+
+
+### Dependencies
+
+* upgrade multiformats to 14 and multiaddr to 13.0.3 ([#3526](https://github.com/libp2p/js-libp2p/issues/3526)) ([5b8813a](https://github.com/libp2p/js-libp2p/commit/5b8813abcbca3fd59e6d137f2a42ff53e115cf99))
+* upgrade to aegir v48 ([#3495](https://github.com/libp2p/js-libp2p/issues/3495)) ([29797a5](https://github.com/libp2p/js-libp2p/commit/29797a5bbbfc5b93d91fcdc244d7f65e2ce5e76c))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.18 to ^5.1.19
+    * @libp2p/echo bumped from ^3.1.5 to ^3.1.6
+    * @libp2p/interface bumped from ^3.2.2 to ^3.2.3
+    * @libp2p/interface-internal bumped from ^3.1.5 to ^3.1.6
+    * @libp2p/logger bumped from ^6.2.7 to ^6.2.8
+    * @libp2p/memory bumped from ^2.0.20 to ^2.0.21
+    * @libp2p/peer-id bumped from ^6.0.9 to ^6.0.10
+    * @libp2p/plaintext bumped from ^3.0.20 to ^3.0.21
+    * @libp2p/utils bumped from ^7.2.1 to ^7.2.2
+    * libp2p bumped from ^3.3.2 to ^3.3.3
+
+## [7.0.23](https://github.com/libp2p/js-libp2p/compare/interface-compliance-tests-v7.0.22...interface-compliance-tests-v7.0.23) (2026-05-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * libp2p bumped from ^3.3.1 to ^3.3.2
+
 ## [7.0.22](https://github.com/libp2p/js-libp2p/compare/interface-compliance-tests-v7.0.21...interface-compliance-tests-v7.0.22) (2026-05-16)