+++ packages/peer-discovery-bootstrap/CHANGELOG.md
@@ -83,6 +83,31 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [12.0.24](https://github.com/libp2p/js-libp2p/compare/bootstrap-v12.0.23...bootstrap-v12.0.24) (2026-05-30)
+
+
+### Dependencies
+
+* upgrade multiformats to 14 and multiaddr to 13.0.3 ([#3526](https://github.com/libp2p/js-libp2p/issues/3526)) ([5b8813a](https://github.com/libp2p/js-libp2p/commit/5b8813abcbca3fd59e6d137f2a42ff53e115cf99))
+* upgrade to aegir v48 ([#3495](https://github.com/libp2p/js-libp2p/issues/3495)) ([29797a5](https://github.com/libp2p/js-libp2p/commit/29797a5bbbfc5b93d91fcdc244d7f65e2ce5e76c))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.2.2 to ^3.2.3
+    * @libp2p/interface-internal bumped from ^3.1.5 to ^3.1.6
+    * @libp2p/peer-id bumped from ^6.0.9 to ^6.0.10
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^7.0.23 to ^7.0.24
+    * @libp2p/logger bumped from ^6.2.7 to ^6.2.8
+
+## [12.0.23](https://github.com/libp2p/js-libp2p/compare/bootstrap-v12.0.22...bootstrap-v12.0.23) (2026-05-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^7.0.22 to ^7.0.23
+
 ## [12.0.22](https://github.com/libp2p/js-libp2p/compare/bootstrap-v12.0.21...bootstrap-v12.0.22) (2026-05-16)